data_pipeline/requirements.txt
watchdog
openai-whisper
eml_parser
nltk
sentence-transformers
chromadb
requests
